Add functionality to route upload to different paths
- Add functionality to read filename of file chosen in `upload`
- Remove the filename extension
- Rename `fetch-result` to `fetch-photo-result`
- Refactor interstitial to use only `fetch-photo-result`
- Depending on filename, route to happy, override, or fatal error journeys
