+++ views/prototype/photo/processing-or-retrieving-image.html
@@ -10,24 +10,16 @@
     {{#nunjucks}}
         {% if values['choose-photo'] == 'upload' %}
             <h3>Processing image</h3>
-            <script>
-                if (localStorage.journey === 'FTC') {
-                    // Error journey
-                    // setTimeout(function(){window.location.href='../photo/upload-errors/not-accepted'},1000);
-                    setTimeout(function(){window.location.href='../photo/questions-intro'},1000);
-                } else {
-                    setTimeout(function(){window.location.href='../photo/questions-intro'},1000);
-                }
-            </script>
         {% elif values['choose-photo'] == 'code' %}
             <h3>Retrieving image</h3>
-            <script>
-                setTimeout(function () {
-                    window.location.href = 'fetch-result'
-                }, 1000);
-            </script>
         {% endif %}
     {{/nunjucks}}
+
+    <script>
+        setTimeout(function () {
+            window.location.href = 'fetch-photo-result'
+        }, 1000);
+    </script>
 </div>
 
 {{/main-content}}
